refactor(st-09085): modularize schema type mapper
