+++ packages/tools/src/data/relational/schema/type-mapper-vendor-maps.ts
@@ -1,0 +1,136 @@
+import type { DatabaseVendor } from '../types.js';
+
+const POSTGRES_TYPE_MAP: Record<string, string> = {
+  smallint: 'number',
+  integer: 'number',
+  int: 'number',
+  int2: 'number',
+  int4: 'number',
+  int8: 'string',
+  bigint: 'string',
+  serial: 'number',
+  bigserial: 'string',
+  smallserial: 'number',
+  real: 'number',
+  float4: 'number',
+  'double precision': 'number',
+  float8: 'number',
+  numeric: 'string',
+  decimal: 'string',
+  money: 'string',
+  text: 'string',
+  'character varying': 'string',
+  varchar: 'string',
+  char: 'string',
+  character: 'string',
+  name: 'string',
+  citext: 'string',
+  boolean: 'boolean',
+  bool: 'boolean',
+  date: 'string',
+  timestamp: 'string',
+  'timestamp with time zone': 'string',
+  'timestamp without time zone': 'string',
+  timestamptz: 'string',
+  time: 'string',
+  'time with time zone': 'string',
+  'time without time zone': 'string',
+  timetz: 'string',
+  interval: 'string',
+  json: 'unknown',
+  jsonb: 'unknown',
+  bytea: 'Buffer',
+  uuid: 'string',
+  inet: 'string',
+  cidr: 'string',
+  macaddr: 'string',
+  macaddr8: 'string',
+  point: 'string',
+  line: 'string',
+  lseg: 'string',
+  box: 'string',
+  path: 'string',
+  polygon: 'string',
+  circle: 'string',
+  xml: 'string',
+  tsvector: 'string',
+  tsquery: 'string',
+  oid: 'number',
+};
+
+const MYSQL_TYPE_MAP: Record<string, string> = {
+  tinyint: 'number',
+  smallint: 'number',
+  mediumint: 'number',
+  int: 'number',
+  integer: 'number',
+  bigint: 'string',
+  float: 'number',
+  double: 'number',
+  'double precision': 'number',
+  decimal: 'string',
+  dec: 'string',
+  numeric: 'string',
+  bit: 'number',
+  char: 'string',
+  varchar: 'string',
+  tinytext: 'string',
+  text: 'string',
+  mediumtext: 'string',
+  longtext: 'string',
+  enum: 'string',
+  set: 'string',
+  boolean: 'boolean',
+  bool: 'boolean',
+  binary: 'Buffer',
+  varbinary: 'Buffer',
+  tinyblob: 'Buffer',
+  blob: 'Buffer',
+  mediumblob: 'Buffer',
+  longblob: 'Buffer',
+  date: 'string',
+  datetime: 'string',
+  timestamp: 'string',
+  time: 'string',
+  year: 'number',
+  json: 'unknown',
+  geometry: 'string',
+  point: 'string',
+  linestring: 'string',
+  polygon: 'string',
+};
+
+const SQLITE_TYPE_MAP: Record<string, string> = {
+  integer: 'number',
+  int: 'number',
+  tinyint: 'number',
+  smallint: 'number',
+  mediumint: 'number',
+  bigint: 'number',
+  real: 'number',
+  double: 'number',
+  'double precision': 'number',
+  float: 'number',
+  numeric: 'number',
+  decimal: 'number',
+  boolean: 'number',
+  text: 'string',
+  varchar: 'string',
+  char: 'string',
+  clob: 'string',
+  'character varying': 'string',
+  'native character': 'string',
+  nchar: 'string',
+  nvarchar: 'string',
+  blob: 'Buffer',
+  date: 'string',
+  datetime: 'string',
+  timestamp: 'string',
+  json: 'unknown',
+};
+
+export const VENDOR_TYPE_MAPS: Record<DatabaseVendor, Record<string, string>> = {
+  postgresql: POSTGRES_TYPE_MAP,
+  mysql: MYSQL_TYPE_MAP,
+  sqlite: SQLITE_TYPE_MAP,
+};
